+++ domestic/models.py
@@ -17,15 +17,7 @@ from wagtail.core.models import Page
 from wagtail.images import get_image_model_string
 from wagtail.images.edit_handlers import ImageChooserPanel
 
-from core import (
-    blocks as core_blocks,
-    cache_keys,
-    cms_slugs,
-    forms as core_forms,
-    helpers,
-    mixins,
-    service_urls,
-)
+from core import blocks as core_blocks, cache_keys, helpers, mixins, service_urls
 from core.blocks import AdvantageBlock
 from core.constants import (
     ARTICLE_TYPES,
@@ -37,9 +29,9 @@ from core.constants import (
 from core.fields import single_struct_block_stream_field_factory
 from core.helpers import build_social_links
 from core.models import CMSGenericPage, Country, IndustryTag, Region, Tag
-from directory_constants import choices
 from domestic import cms_panels, forms as domestic_forms
 from domestic.helpers import build_route_context, get_lesson_completion_status
+from exportplan.core import helpers as exportplan_helpers
 
 
 class DataLayerMixin(
@@ -175,17 +167,12 @@ class DomesticDashboard(
 
     def get_context(self, request):
         user = request.user
-
         context = super().get_context(request)
         context['visited_already'] = user.has_visited_page(self.slug)
         user.set_page_view(self.slug)
-        context['export_plan_progress_form'] = core_forms.ExportPlanForm(
-            initial={'step_a': True, 'step_b': True, 'step_c': True}
-        )
-        context['industry_options'] = [{'value': key, 'label': label} for key, label in choices.SECTORS]
+        context['exportplan_list'] = exportplan_helpers.get_exportplan_detail_list(user.session_id)
         context['export_opportunities'] = helpers.get_dashboard_export_opportunities(user.session_id, user.company)
         context.update(get_lesson_completion_status(user, context))
-        context['export_plan_in_progress'] = user.has_visited_page(cms_slugs.EXPORT_PLAN_DASHBOARD_URL)
         context['routes'] = build_route_context(user, context)
         return context