+++ react-components/src/components/ProductFinder/ClassificationTree.test.jsx
@@ -93,7 +93,7 @@ describe('Classification tree', () => {
     })
     let arrows = container.querySelectorAll('.classification-tree__arrow')
     expect(arrows.length).toEqual(2)
-    let levels = container.querySelectorAll('.classification-tree__item > span')
+    let levels = container.querySelectorAll('.classification-tree__item > div')
     expect(levels.length).toEqual(3)
     expect(levels[2].textContent).toMatch(
       'Processed cheese, not grated or powdered'